[MOV] filter: rename fitler_icons_overlay file
Rename fitler_icons_overlay to filter_icons_overlay

X-original-commit: 752f2f208cc8aeb9185e4725377299d3b86623be
