+++ src/components/filters/filter_icons_overlay/filter_icons_overlay.xml
